feat: link login

+++ frontend/src/pages/login/index.jsx
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React, { useState, useEffect } from "react";
 import peach_logo from "../../../public/peach_logo.png";
 import kakao from "../../images/kakao.png";
 import google from "../../images/google.png";
@@ -6,10 +6,22 @@ import naver from "../../images/naver.png";
 import Image from "next/image";
 import BasicBtn from "@/components/button/BasicBtn";
 import Link from "next/link";
-const index = () => {
+
+const Index = () => {
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
   const [message, setMessage] = useState("");
+  const [rememberMe, setRememberMe] = useState(false);
+  const [keepLoggedIn, setKeepLoggedIn] = useState(false);
+
+  useEffect(() => {
+    const savedEmail = localStorage.getItem("savedEmail");
+    if (savedEmail) {
+      setEmail(savedEmail);
+      setRememberMe(true);
+    }
+  }, []);
+
   const handleLogin = async (e) => {
     e.preventDefault();
 
@@ -28,7 +40,19 @@ const index = () => {
     console.log(data);
     if (response.ok) {
       setMessage(data.message);
-      sessionStorage.setItem("token", data.token); // sessionStorage에 토큰 저장
+
+      if (keepLoggedIn) {
+        localStorage.setItem("token", data.token); // 로그인 유지
+      } else {
+        sessionStorage.setItem("token", data.token); // 세션 스토리지에 토큰 저장
+      }
+
+      if (rememberMe) {
+        localStorage.setItem("savedEmail", email); // 아이디 저장
+      } else {
+        localStorage.removeItem("savedEmail");
+      }
+
       window.location.href = "/";
     } else {
       setMessage(data.message);
@@ -37,26 +61,38 @@ const index = () => {
 
   return (
     <div className="center1">
-      <Image src={peach_logo} width={300} />
-      <form className="w-1/5 mt-10 min-w-60 " onSubmit={handleLogin}>
+      <Image src={peach_logo} width={300} alt="Peach Logo" />
+      <form className="w-1/5 mt-10 min-w-60" onSubmit={handleLogin}>
         <div className="flex items-center">
-          <input type="checkbox" className="w-5  h-5 text-gray form-checkbox" />
+          {keepLoggedIn}
+          <input
+            type="checkbox"
+            className="w-5 h-5 text-gray form-checkbox"
+            checked={keepLoggedIn}
+            onChange={(e) => setKeepLoggedIn(e.target.checked)}
+          />
           <span className="ml-2 text-sm mr-3">로그인 유지</span>
-          <input type="checkbox" className="w-5 h-5 text-gray form-checkbox" />
+          <input
+            type="checkbox"
+            className="w-5 h-5 text-gray form-checkbox"
+            checked={rememberMe}
+            onChange={(e) => setRememberMe(e.target.checked)}
+          />
           <span className="ml-2 text-sm">아이디 저장</span>
         </div>
-        <div className="mt-2 flex items-center  w-full py-3 bg-[#f8f8f8] border-[1px] border-solid border-[#808080]">
+        <div className="mt-2 flex items-center w-full py-3 bg-[#f8f8f8] border-[1px] border-solid border-[#808080]">
           <input
             type="text"
-            className=" bg-[#f8f8f8] ml-3 text-[20px] outline-none "
+            className="bg-[#f8f8f8] ml-3 text-[20px] outline-none"
             placeholder="아이디"
+            value={email}
             onChange={(e) => setEmail(e.target.value)}
           />
         </div>
         <div className="my-5 w-full flex items-center py-3 bg-[#f8f8f8] border-[1px] border-solid border-[#808080]">
           <input
             type="password"
-            className=" bg-[#f8f8f8] ml-3 text-[20px] outline-none  "
+            className="bg-[#f8f8f8] ml-3 text-[20px] outline-none"
             placeholder="비밀번호"
             onChange={(e) => setPassword(e.target.value)}
           />
@@ -65,7 +101,6 @@ const index = () => {
 
         <BasicBtn
           text={"로그인"}
-          // onClick={handleLogin}
           type="submit"
           bgColor={"#fb5e67"}
           textColor={"#fff"}
@@ -79,20 +114,20 @@ const index = () => {
       </form>
       <div className="m-4 center1">
         <div className="flex items-center min-w-60 w-full">
-          <div className=" w-2/12  border-[1px]  border-[#808080]"></div>
+          <div className="w-2/12 border-[1px] border-[#808080]"></div>
           <div className="text-gray-500 text-center text-sm w-10/12">
             소셜 계정으로 간편 로그인
           </div>
-          <div className=" w-2/12 border-[1px]  border-[#808080]"></div>
+          <div className="w-2/12 border-[1px] border-[#808080]"></div>
         </div>
         <div className="flex">
-          <Image src={kakao} width={40} className="m-3" />
-          <Image src={google} width={40} className="m-3" />
-          <Image src={naver} width={40} className="m-3" />
+          <Image src={kakao} width={40} className="m-3" alt="Kakao" />
+          <Image src={google} width={40} className="m-3" alt="Google" />
+          <Image src={naver} width={40} className="m-3" alt="Naver" />
         </div>
       </div>
     </div>
   );
 };
 
-export default index;
+export default Index;
